MemoryUtil native integer size update + tests

+++ lib/Utils/src/Memory/IUnmangedHeap.cs
@@ -38,7 +38,7 @@ namespace VNLib.Utils.Memory
         /// <param name="elements">The number of elements to allocate</param>
         /// <param name="zero">An optional parameter to zero the block of memory</param>
         /// <returns></returns>
-        IntPtr Alloc(UInt64 elements, UInt64 size, bool zero);
+        IntPtr Alloc(nuint elements, nuint size, bool zero);
 
         /// <summary>
         /// Resizes the allocated block of memory to the new size
@@ -47,7 +47,7 @@ namespace VNLib.Utils.Memory
         /// <param name="elements">The new number of elements</param>
         /// <param name="size">The size (in bytes) of the type</param>
         /// <param name="zero">An optional parameter to zero the block of memory</param>
-        void Resize(ref IntPtr block, UInt64 elements, UInt64 size, bool zero);
+        void Resize(ref IntPtr block, nuint elements, nuint size, bool zero);
 
         /// <summary>
         /// Free's a previously allocated block of memory
